+++ lib/objects/animated_object.dart
@@ -25,7 +25,7 @@ class AnimatedObject extends GameComponent {
 
   @override
   void update(double dt) {
-    animation?.update(dt);
     super.update(dt);
+    animation?.update(dt);
   }
 }